Move Conway specific instances from Core to Conway/Base
